requirements.txt
hydra-core
black
torch
tqdm
dm-tree
Pillow
opencv-python
matplotlib
imageio
moviepy
einops
einops_exts
tokenizers
pytorch-lightning==1.9.4
transformers
wandb
rich
gputil
ipdb
gdown
seaborn
tabulate